Updates for portability.

+++ shapes/Plane.cpp
@@ -45,7 +45,7 @@ void Plane::Create(PrimitiveList & primitiveList, const unify::Parameters & para
 		BufferSet & set = primitiveList.AddBufferSet();
 		set.SetEffect( effect );
 
-		set.AddMethod( RenderMethod::CreateTriangleListIndexed( vertexCount, indexCount, 0, 0 ) );
+		set.AddMethod( RenderMethod::CreateTriangleListIndexed( vertexCount, (unsigned int)indexCount, 0, 0 ) );
 
 		unsigned short stream = 0;
